+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### v0.1.34
 - EffectDisplay.`DrawPriority` : `PAST` -> `POST`
-- Callback : Add `Callback.ON_SHIELD_BREAK`
+- Callback
+    - Add `Callback.ON_SHIELD_BREAK`
+    - Add `Callback.ON_SKILL_ACTIVATE`
 
 ### v0.1.33
 - Actor : Fix `fire_*` methods not accepting Sprite wrappers